Arraylist
Java program that takes 15 values of type integer as inputs from user, store the values in an array. ) Ask user to enter a number, check if that number (entered by user) is present in array
or not. If it is present print, “the number found at index (index of the number) ” and the text “number not found in this array”